// asr.webapp.v2/src/app/modules/financial/statement/statement.types.ts
import { Paging } from "app/core/type/paging/paging.type";
import { BaseModel } from "app/core/type/base/BaseModel";

export class StatementDetail extends BaseModel {
    logId: number;
    financialResourceDetailId: number;
    financialResourceId: number;
    financialResourceName: string;
    financialResourceTypeId: number;
    financialResourceTypeName: string;
    ownerId: number;
    ownerName: string;
    ownerTypeId: number;
    ownerTypeName: string;
    financialTransactionTypeId: number;
    financialTransactionTypeName: string;
    transactionTypeColor: string;
    bankId: number;
    bankName: string;
    url: string;
    accountHolder: string;
    currencyId: number;
    currencyName: string;
    currencyAbbreviation: string;
    transactionLocalDate: string;
    chequeId: number;
    chequeNumber: string;
    chequeLocalDate: string;
    amount: number;
    balance: number;
    paymentTypeId: number;
    paymentTypeName: string;
    requestId: number;
    requestDetailId: number;
    requestTypeId: number;
    requestTypeName: string;
    statementDescription: string;
    financialTransactionTypeIdList: Array<Number>;
    partitionKeyList: Array<Number>;
    transactionLocalDateFrom: string;
    transactionLocalDateTo: string;
    amountFrom: number;
    amountTo: number;
    balanceFrom: number;
    balanceTo: number;
    status: number;
    credit: boolean;
    ownerIdList: Array<Number>;
    page: Paging;

}
export class Statement extends BaseModel {
    statementId: number;
    financialResourceId: number;
    financialResourceName: string;
    financialResourceTypeId: number;
    financialResourceTypeName: string;
    ownerId: number;
    ownerName: string;
    ownerTypeId: number;
    ownerTypeName: string;
    currencyId: number;
    currencyName: string;
    currencyAbbreviation: string;
    chequeId: number;
    chequeNumber: string;
    bankId: number;
    bankName: string;
    accountHolder: string;
    bankIdList: Array<number>;
    url: string;
    chequeDate: string;
    chequeLocalDate: string;
    chequeLocalDateClean: string;
    balance: number;
    paymentTypeId: number;
    paymentTypeName: string;
    paymentTypeIdList: Array<number>;
    status: number;
    statusDescription: string;
    statusColor: string;
    balanceFrom: number;
    balanceTo: number;
    credit: boolean;
    ownerIdList: Array<Number>;
    page: Paging;
}